Add files via upload
Jupyter Notebook that simulates a sampling distribution for a sample mean from a population of coins--pennies, nickels, dimes, and quarters. You can download this and run it in colab.google.com or Jupyter Lab at jupyter.org. You do not need to create an account or sign in to either of these sites to run this Notebook. Try it and let me know what you think!